+++ src/containers/RegionsContainer.test.jsx
@@ -6,7 +6,7 @@ import { useDispatch, useSelector } from 'react-redux';
 
 import RegionsContainer from './RegionsContainer';
 
-import REGIONS from '../fixtures/regions';
+import REGIONS from '../../fixtures/regions';
 
 describe('RegionsContainer', () => {
   const dispatch = jest.fn();
