+++ src/containers/RestaurantsContainer.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { useSelector } from 'react-redux';
 
-import { get } from './utils';
+import { get } from '../utils';
 
 export default function RestaurantsContainer() {
   const restaurants = useSelector(get('restaurants'));
