+++ src/pages/NotFoundPage.jsx
@@ -4,6 +4,7 @@ export default function NotFoundPage() {
   return (
     <div>
       <p>잘못된 페이지입니다</p>
+      <p>404 Not Found</p>
     </div>
   );
 }
