+++ src/pages/RestaurantsPage.jsx
@@ -2,9 +2,9 @@ import React, { useEffect } from 'react';
 
 import { useDispatch } from 'react-redux';
 
-import RegionsContainer from '../RegionsContainer';
-import CategoriesContainer from '../CategoriesContainer';
-import RestaurantsContainer from '../RestaurantsContainer';
+import RegionsContainer from '../containers/RegionsContainer';
+import CategoriesContainer from '../containers/CategoriesContainer';
+import RestaurantsContainer from '../containers/RestaurantsContainer';
 
 import {
   loadInitialData,
